backend/utils.py
from werkzeug.security import generate_password_hash, check_password_hash

def hash_password(raw: str) -> str:
    return generate_password_hash(raw)

def verify_password(pw_hash: str, raw: str) -> bool:
    return check_password_hash(pw_hash, raw)
