requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
marshmallow==3.21.3
python-dotenv==1.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
pytest==8.3.3
requests==2.32.3
